requirements.txt
Flask==2.0.2
Pillow==8.4.0
pdfplumber==0.5.28
img2pdf==0.4.3
opencv-python==4.5.4.60
